Lift out the thread-pool into its own non-global obj and package.
The thread-pool used to be a global object, but now it lives in a slot
in the CORE. While currently each time a CORE gets created it will come
with its own thread-pool, it is not finalized if it is going to stay like
this. Some code around how the CORE get initialized and how the engine
is initialized/deinitialized changed as a result of this. The
threading model or how we used the threads hasn't changed.

The thread-pool now lives in the virality.thread-pool package with a
nickname of tpool. All code has been fixed to honor this change.

I did a slight change of how *core-debug* works in that now it is always
bound to something: either *no-core-value* or an actual core. This was to
remove compiler errors when *core-debug* was truly unbound (which used to
be ok) when during macro expansion the symbol was atually being checked.
Now the *core-debug* symbol must always be bound and we use a sentinel
to indicate if it is currently has a CORE bound do it or not.

I fixed a small situation where the profiling code was using the wrong
libraries. I couldn't separate this in magit, sorry.