Checkpoint commit of initial implementation of resource-cache.
The initial implementation is almost done. I did this commit cause it was
a bunch of code that I wanted safe in github while I worked on it.

This code is intending to replace the original resource-cache which
suffers from both an architectural layering problem and also an inverted
control problem which makes it very difficult to plan and concurrently
load things into the cache. This new implementation is much better
designed and easier to put into a layer structure in the engine.