Continuing the design of the new resource-cache and friends API.
This part details the CACHE-ITEM class which is often the base class of a value
you want to put into the resource-cache. It also details the warming protocol
and most of the scheduler and executor classes and protocols. The warming
protocol is how keys are (possibly) concurrently changed into values and put
into the resource-cache.