Checkpoint of work on psilord/issue-40 branch.
There is enough work here that I am frightened of loss if my machine blows up
and my backups fail, etc, etc. So I'm committing it to github and will continue
the work.

This implements almost all of the parser for the texture-map DSL. The only
thing left for the DSL is cube-maps when their faces are combined into a single
image and a generic function typechecker for texture map AST trees.  There is a
fair bit of TODO documentation spread around the code which I will get to as I
finish things.

The next major steps are reifying the texture-map =meta/texture-map= data into
the CORE of the engine, then fixing how the texture-map data is concurrently
loaded/generated and how the texture DSL uses and references the data.