Fixed up the attribute-bag base class so it works.
I need this for the various DSLs in V such as define-texture-map,
define-texture, and define-material. It can surely be used in more places
too. define-texture and define-material use ad hoc attr/value stores
inside of themselves and they need to be refactored to use this one.
define-texture-map will start out by using this one.