Texture-map implementation uses abags. Refactor of parse algorithm.
Abags are attribute-bags which are used as base classes in the texture-map ast
representation.

Factored some more of the parse-tree control-flow of the texture-map DSL
processor into a traditional LL() style compiler pass.

Now when the parser is finished, the generated code collates any runtime
created texture-map ASTs (to support cube maps) so we can handle them in one
batch when reifying into core.

The design of the parser is an extensible compiler where you specialize
different methods that perform the parsing in accordance to the context of
the parsing. Seems to work pretty well, actually. Since there were a pile
of different kinds of texture-maps (and an unbounded direction with how
cube maps can be specified!) this seemed a logical thing to do. The code
ended up pretty short for all the functionality it has.

Needs more testing, but this is far enough it is worth a commit.