Added many doctrings to attribute-bag code. Added optional args.
It used to be the case that attempting to set a computed attr for an
attribute-value that didn't exist would be an error. Now you can use an
optional argument to supply a semantic value in this context and just let the
new attribute-value be inserted. If you don't supply the optional
semantic-value, it is still an error (to try and set only a computed value for
a non-existent attribute-value).