Checkpoint commit of the CLONE and CLONE-OBJECT code.
This code is still somewhat experimental. It's job is to provide the ability
to shallow, deep, or any other kind of clone an object, list, tree, etc.
However, other than for cons cells, hash tables, and arrays, it is not
generic. Users of this API must generally write a CLONE and CLONE-OBJECT
method for their types.

I changed a single place in the attribute-bag to use the CLONE code for
testing.

This commit contains about half of a test suite for the clone code, too.

The closing code does need, before it can be considered done, the ability
to clone up to EQL comparison for things like cons cells and whatnot.
I suspect I will have to add additional clone-policy specializations to
distinguish those situations.