Partially changed book keeping of EQL-MAP data structure to be explicit.
This is a checkpoint commit. I need to finish this work.

While adding the EQL-MAP, which is a data structure to allow a cloning to
recover complex referential structure, I discovered that it can't be a
special and it can't be in a shared slot, because it was not stable in the
face of concurrency. So it is explicitly passed to each CLONE or CLONE-OBJECT
method.

I'm also adding in the cons/list/alist/tree variants for shallow cloning.