Reworked intentions & added comparisons of them.
The intentions got separated out so shallow-clone-cons became two
objects: shallow-clone and cons-intention. Fixing up the CLONE and
CLONE-OBJECT API in order to deal with this change.

This change happened because I ran into a shared structure situation where
a cons cell was used as an alist pair, but also as list structure in another
list. While I'm not going to solve that problem here, this rework allows us
to observe it and do something about it when we do want to solve it.

Move the clone testing code to a test file and make some of the tests more
detailed.