Checkpoint commit of cloning code.
Added statistics to the eql-map so we can unit test the exact number of
reference/value moves and allocations we expect in our test cases.  Now, under
a "domain" of which there are currently two: :move and :allocation, statistics
are kept about how many events of what exact type fell under each type.  A
:move is when a setf of a reference/value happens from the old structure to the
new one. And an :allocation happens when new memory is actually constructed to
clone an object.

So, in cloning a cons cell (cons 1 #\a) with a cons-intention, the
:allocation domain will have 1 CONS type allocated, and in the :move
domain there will be 1 BIT (or thereabouts) and 1 STANDARD-CHAR (or
thereabouts) whose reference/value was moved into the allocated cons cell.
Basically :move represents the edge of the allocating and where objects
start getting shared from the original structure.

Added additional API to eql-map to manage the statistics and additional
load on the test functions.

Added a function to check the statistics for each domain to ensure they
are exactly what we desire them to be.

I'm not yet done pushing this change through the code, but it was getting
to be a large and I was getting ansty to only have it on my machine.