Checkpoint commit. Added ALLOCATABLEP generic function.
Reworked statistics and added a new domain :move-clone which records when we
moved a previously cloned value to a new spot. Made the eql-map ONLY contain
allocated stuff with transitions in it and now ALL non allocatable things are
considered just moves in the :move-original domain. This lowers the accuracy a
little bit (but in some places increases it) of the statistics, but increases
speed greatly (with huge memory reduction) when copying large arrays of
single-floats and things like that.

Changed many (but not yet all) of the unit tests to accomodate the changed
statistics.