Added unimplemented INTENTION classes and notes for the clone system
The *-pedantic intentions indicate that when cyclic structure is found for
those types of structure to NOT honor it. This is not the default and is very
surprising in a lisp context. Currently it is unimplemented.  If we want to
implement it, then we'd make a CLONE-SHALLOW-LIST-PEDANTIC and so forth set of
function calls, and then implement CLONE-OBJECT which will be very similar to
the non pedantic ones but it will not preserve the cyclic strucure.