Updated a clone unit test to be more rigorous.
It was the test-clone-deep-new-type-%foo/%bar-0 test. It used to just use
numbers, but that was too easy. I fixed it to the base class as a list in its
slot and the derived class has a reference to that list. Then after a deep
clone, the structure should be perfectly preserved. I added a pile of asserts
to ensure it is ok for the unit test.