Change the clone system to explicitly whitelist allowed types.
+ The whitelisting allows us to catch when a user of the clone interface
attempts to clone a type the cloning system doesn't know about. Added a nice
no-applicable-method error that explains what happened and likely how to fix
it. The behavior before this change was to silently identity clone the unknown
object. When using the old behavior with a collection of types I had just made,
I got very surprised because I forgot to add one of the types into the cloning
system and it identity cloned it. I would much rather have had it signal a
condition letting me know it doesn't know what to do. Now it does.

+ Add :argument-precedence-order to the CLONE generic function.  This allows us
to prioritize the policy and intention when cloning the object without having
to have them first in the argument sequence.