Add a default CLONE-ALLOCATE for STANDARD-CLASS.
This will call make-instance to create a new instance of the object given its
TYPE-OF. This generally removes having to define this method for individual
classes unless clone-allocate need to do something special.

We use make-instance instead of allocate-instance because we generally desire
the initargs and initforms to have been honored in the cloning system.