Checkpoint commit. Redesign of the texture-map internals and DSL.
The redesign ocurred because the programmatic API was too weak in the original
design and when it came time to integrate the texture-map codes into the
concurrent loading of the resources--it failed to represent the domain
properly. This new design has now a very strong API that both the engine and
the appdev may use. Also the texture-map DSL was redisgned in order for
extensibility and understandability reasons.

The code will load and not break, but the texture-map feature is still not
integrated into the engine.