Removed previous implementation of texture-map-dsl parser.
Changed the texture-map-descriptor to match the new methodology.  The
texture-map table should do the right thing by calling the constructor for the
programmatic API created by the texture-map parser which is executed at macro
expansion time. The meta table for texture-maps should be able to deal with
them when they populate.

The define-texture-map macro is still not functional, but will be soon.