Changed define-texture-map to new DSL syntax. Fixed test suite.
Currently DEFINE-TEXTURE-MAP is still exported in package C and the code uses
it that way, but we'll likely change it to the TEXMAP package to match all the
syntax datum symbols for that DSL (and start depreciating or changing the use
of the C package). The metaspace processing of the texture-map DSL seems
reasonably complete--but there will be some minor changing in the future. This
code sort of checkpoints this state before the runtime gets worked on.

The test suite compiles, loads, and executes again!

The test suite for DEFINE-TEXTURE-MAP is wholly incomplete. This will likely
get fixed in a near future commit.