Fixup resource-cache warming protocol API to match actual needs.
The warming pprotocol now has a complete asynchronous API so we can
load data across frames or otherwise incrementally.

Add a state object to the texture-map type to make it easier to track
the stages of loading and validity for a texture-map.

Add the files and initial sketch of the code which will materialize and
realize the texture-maps into memory and the gpu.

Make some comments about where the next chunks of code should go.