Connected resource-cache warming-protocol to texture-map materialization.
There are exhaustive tests, but I tested by hand. It seems to work in that we
can materialize the data-elements of a texture-map into the resource cache.

Next step is rectification of the texture-map in-memory data structure.
Rectification fills in any missing data that we must get from the actual
images, and asserts that whatever constraints the dsl form requested are
honorable.