Fixed: texture-map warming protocol, classify-rectification, added rectify.
Completed classification of rectification into :synthesize, :validate, :infer
for all texture-maps variants.

The entry point to rectification has been fixed, but the recursive
descent in order to :synthesize or :validate the texture-map is not
yet written.

Deduction of number of needed mipmaps is still a bit screwed up.

Added more stuff to the texture-map state object.

This is a checkpoint commit.