Changed list.lisp, which used VUTILS:DEFAULT to not use it.
This is because of a load ordering problem where list.lisp got compiled
before macros.lisp and hence DEFAULT was though to be a function when it
finally got overwritten by a macro--leading to a warning during compilation.

Instead of restructuring vutils better, I jut reverted this use of DEFAULT
and coded it the traditional way. This fixes the bug without the stumble
on getting bigger. Later, we'll inspect it and rego the deppendencies a little
better.